% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readheftydata.R
\name{findline}
\alias{findline}
\title{Find Line With String}
\usage{
findline(fn, lnstr)
}
\arguments{
\item{fn}{a filename of HeFTy text file output}

\item{lnstr}{string to search for}
}
\value{
line number
}
\description{
Find Line With String
}
