% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_eU.R
\name{plot_date_eu}
\alias{plot_date_eu}
\title{A plotting function for  (U-Th)/He data}
\usage{
plot_date_eu(hedf, bestfitdf = NULL, eUbin = NULL, facet_vars = "Sample")
}
\arguments{
\item{hedf}{tidy dataframe with columns labeled "eU", "Date", "Unc","Color", and "Sample"}

\item{bestfitdf}{dataframe containing prediction of best fit path from hefty. must have column names equal to "eU" and "Date"}

\item{eUbin}{value or vector for eU value(s) where bins are split for HeFTy modeling}

\item{facet_vars}{which variable to facet on, defaults to `Sample`. Currently the only option}
}
\description{
This function will take a tidy dataframe of (U-Th)/He data and make simple plots it
}
