% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readheftydata.R
\name{readconstraint}
\alias{readconstraint}
\title{Read Single Hefty Constraint}
\usage{
readconstraint(fn, smpl)
}
\arguments{
\item{fn}{a filename of HeFTy text file output}

\item{smpl}{sample name for hefty file}
}
\value{
dataframe with constraints
}
\description{
This reads the hefty constraints and assigns the column names that will be used for plotting
}
