% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readheftydata.R
\name{readconstraints}
\alias{readconstraints}
\title{Read Hefty Constraints}
\usage{
readconstraints(fn, smpl)
}
\arguments{
\item{fn}{a filename or vector of filenames of HeFTy text file output}

\item{smpl}{sample name or vector of sample names for a for hefty file}
}
\value{
dataframe with constraints
}
\description{
Read Hefty Constraints
}
\examples{
fn=system.file("extdata","HeFTyOut-Sample1.txt",package="thermochronplotr")
readconstraints(fn)
# NOTE: fn should point to your own file on your system
}
