% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readheftydata.R
\name{readpath}
\alias{readpath}
\title{Extract Single Model Time Temp paths}
\usage{
readpath(fn, smpl)
}
\arguments{
\item{fn}{filename of hefty output}

\item{smpl}{sample name for hefty file}
}
\value{
dataframe of paths
}
\description{
Extract Single Model Time Temp paths
}
