% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readheftydata.R
\name{readpaths}
\alias{readpaths}
\title{Extract Time Temp paths}
\usage{
readpaths(fn, smpl)
}
\arguments{
\item{fn}{filename or vector of filenames of hefty output}

\item{smpl}{sample name or vector of sample names for hefty file}
}
\value{
dataframe of paths
}
\description{
Extract Time Temp paths
}
