% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_plot.R
\name{shuffle_plots}
\alias{shuffle_plots}
\title{A function to plot HeFTy and date-eU plots side by side (shuffled)}
\usage{
shuffle_plots(
  hedf,
  bestfitdf = NULL,
  heftydf,
  constraints,
  synthetic = NULL,
  eUbin = NULL,
  helim = c(NA, NA),
  eUlim = c(NA, NA),
  timelim = c(NA, NA),
  templim = c(NA, NA)
)
}
\arguments{
\item{hedf}{tidy dataframe with (U-Th)/He data and columns labeled "eU", "Date", "Unc", "Color", "Sample", "Description", "Elevation_m"}

\item{bestfitdf}{dataframe containing prediction of best fit path from hefty. must have column names equal to "eU", "Date",and "Sample"}

\item{heftydf}{tidy dataframe extracted from HeFTy paths extracted with `readpaths()`}

\item{constraints}{dataframe with HeFTy constraints extracted with `readconstraints()`}

\item{synthetic}{dataframe with synthetic/binned data points used in HeFTy modelling. Must have column names equal to "eU", "Date", "Unc", and "Sample"}

\item{eUbin}{value or vector for eU value(s) where bins are split for HeFTy modeling}

\item{helim}{limits for the y (He date) axis, must be a vector of 2-defaults to NA where limits are set based on the total dataset starting at 0}

\item{eUlim}{limits for the x (eU) axis, must be a vector of 2-defaults to NA where limits are set based on the total dataset starting at 0}

\item{timelim}{vector of length 2 with limits for time (x) axis. Defaults to NA where limits are set based on the total dataset starting at 0}

\item{templim}{vector of length 2 with limits for temperature (y) axis. Defaults to NA, where limits are set based on the total dataset}
}
\description{
This function will take a tidy dataframe of (U-Th)/He data and make simple plots it
}
\examples{
fn=system.file("extdata","AHe_data.csv",package="thermochronplotr")
hedf<-readr::read_csv(fn)
fn2=system.file("extdata","ForwardModels.csv",package="thermochronplotr")
bestfitdf <- readr::read_csv(fn2)
fn3 = system.file("extdata","HeFTyOut-Sample1",package="thermochronplotr")
fn4 = system.file("extdata","HeFTyOut-Sample2",package="thermochronplotr")
heftydf<-readpaths(c(fn3,fn4),c('Sample1','Sample2'))
constraints<-readconstraints(c(fn3,fn4),c('Sample1','Sample2'))
shuffle_plots(ahe,forward,hefty,constraints,eUbin = 15)
}
