% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hefty_output.R
\name{plot_hefty_output}
\alias{plot_hefty_output}
\title{A plotting function for single HeFTy inversions}
\usage{
plot_hefty_output(dF, constraints, facet_vars = "Sample")
}
\arguments{
\item{dF}{tidy dataframe extracted from HeFTy paths extracted with `readpaths()`}

\item{constraints}{dataframe with HeFTy constraints extracted with `readconstraints()`}

\item{facet_vars}{which variable to facet on, defaults to `Sample`. Currently the only option}
}
\description{
This function will take a tidy dataframe made from the text file output of a HeFTy inversion and plot it
}
\examples{
HeFTySinglePlot()
}
